Pull out the compiler and related dependencies.
