+++ lib/errors/builderror.js
@@ -1,8 +1,8 @@
 var colors = require("colors");
-var ExtendableError = require("./extendableerror");
+var TruffleError = require("truffle-error");
 var inherits = require("util").inherits;
 
-inherits(BuildError, ExtendableError);
+inherits(BuildError, TruffleError);
 
 function BuildError(message) {
   message = "Error building:\n\n" + message + "\n\n" + colors.red("Build failed. See above.");
