+++ lib/errors/taskerror.js
@@ -1,7 +1,7 @@
-var ExtendableError = require("./extendableerror");
+var TruffleError = require("truffle-error");
 var inherits = require("util").inherits;
 
-inherits(TaskError, ExtendableError);
+inherits(TaskError, TruffleError);
 
 function TaskError(message) {
   TaskError.super_.call(this, message);
