+++ lib/test.js
@@ -7,12 +7,13 @@ var fs = require("fs");
 var Web3 = require("web3");
 var Config = require("./config");
 var Contracts = require("./contracts");
-var Profiler = require("./profiler");
 var TestRunner = require('./testing/testrunner');
 var TestSource = require("./testing/testsource");
 var Deployed = require('./testing/deployed');
 var expect = require("truffle-expect");
 var provision = require("truffle-provisioner");
+var find_contracts = require("truffle-contract-sources");
+var SolidityUtils = require("truffle-solidity-utils");
 var async = require("async");
 
 chai.use(require("./assertions"));
@@ -92,8 +93,8 @@ var Test = {
       // compile what's necessary.
       function(c) {
         async.parallel({
-          contract_files: Profiler.all_contracts.bind(Profiler, options.contracts_directory),
-          test_files: Profiler.all_contracts.bind(Profiler, options.test_directory)
+          contract_files: find_contracts.bind(find_contracts, options.contracts_directory),
+          test_files: find_contracts.bind(find_contracts, options.test_directory)
         }, function(err, result) {
           if (err) return c(err);
 
@@ -140,7 +141,7 @@ var Test = {
         async.each(testContracts, function(contract, finished) {
           var file = file_hash[contract.contract_name];
 
-          Profiler.ordered_abi(file, contract, function(err, ordered) {
+          SolidityUtils.ordered_abi(file, contract.abi, contract.contract_name, function(err, ordered) {
             if (err) return finished(err);
 
             contract.abi = ordered;
