+++ lib/testing/deployed.js
@@ -1,5 +1,3 @@
-var Contracts = require("../contracts");
-var Profiler = require("../profiler");
 var path = require("path");
 
 var Deployed = {
