+++ package.json
@@ -25,11 +25,16 @@
     "spawn-args": "^0.1.0",
     "temp": "^0.8.3",
     "truffle-artifactor": "^1.0.2",
+    "truffle-compile": "0.0.1",
     "truffle-contract": "^1.0.2",
+    "truffle-contract-sources": "0.0.1",
     "truffle-deployer": "0.0.1",
+    "truffle-error": "0.0.1",
     "truffle-expect": "0.0.3",
     "truffle-provisioner": "0.0.1",
     "truffle-require": "0.0.1",
+    "truffle-resolver": "0.0.1",
+    "truffle-solidity-utils": "0.0.1",
     "web3": "^0.16.0",
     "yargs": "^3.27.0"
   },
